+++ packages/layer-manager/src/components/withRenderTarget.js
@@ -4,8 +4,13 @@ import PropTypes from 'prop-types';
 
 import { Gateway, GatewayRegistry } from './gateway';
 import Portal from './Portal';
-// eslint-disable-next-line
-export default function withRenderTarget({ target }: Event, WrappedComponent: ReactClass<*>) { 
+import type { ComponentType } from '../types';
+
+export default function withRenderTarget(
+  { target, wrapWithTransitionGroup }:
+  { target: string, wrapWithTransitionGroup: boolean },
+  WrappedComponent: ComponentType
+) {
   return class extends Component {
     static contextTypes = {
       gatewayRegistry: PropTypes.instanceOf(GatewayRegistry),
@@ -16,7 +21,7 @@ export default function withRenderTarget({ target }: Event, WrappedComponent: Re
       const GatewayOrPortal = gatewayRegistry ? Gateway : Portal;
 
       return (
-        <GatewayOrPortal into={target}>
+        <GatewayOrPortal into={target} wrapWithTransitionGroup={wrapWithTransitionGroup}>
           <WrappedComponent {...this.props} />
         </GatewayOrPortal>
       );