+++ packages/spotlight/stories/00-basic-story.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { storiesOf } from '@kadira/storybook';
+
+import { name } from '../package.json';
+import BasicExample from './examples/basic';
+
+storiesOf(name, module)
+  .add('Basic Usage', () => <BasicExample />);